Refactor base class into utils and read/write class.
