+++ src/barometerdrivers/__init__.py
@@ -1,2 +1,2 @@
-from .hp206c import HP206C       # noqa: F401
-from .ms5803 import MS5803_01BA  # noqa: F401
+# from .hp206c import HP206C       # noqa: F401
+# from .ms5803 import MS5803_01BA  # noqa: F401
